refactor(backend/middleware): Abstract verification functions into utils for readability
